# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TAMS AI Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /decisions:
    get:
      operationId: listDecisions
      tags: [decisions]
      summary: List decisions
      responses:
        "200":
          description: Decisions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Decision"
    post:
      operationId: createDecision
      tags: [decisions]
      summary: Create a decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionInput"
      responses:
        "201":
          description: Created decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
  /decisions/{id}:
    get:
      operationId: getDecision
      tags: [decisions]
      summary: Get a decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
    patch:
      operationId: updateDecision
      tags: [decisions]
      summary: Update a decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionUpdate"
      responses:
        "200":
          description: Updated decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
    delete:
      operationId: deleteDecision
      tags: [decisions]
      summary: Delete a decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /decisions/{id}/analyze:
    post:
      operationId: analyzeDecision
      tags: [decisions]
      summary: Run AI + Red Team analysis on a decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated decision with AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
  /decisions/{id}/tasks:
    post:
      operationId: createTasksFromDecision
      tags: [decisions]
      summary: Create tasks from a decision (turn decision into action items)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TasksFromDecisionInput"
      responses:
        "201":
          description: Created tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  # ── ASSETS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, done, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Decision:
      type: object
      required: [id, title, status, confidenceScore, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        context:
          type: ["string", "null"]
        options:
          type: array
          items:
            type: string
        advantages:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        aiAdvice:
          type: ["string", "null"]
        redTeamAdvice:
          type: ["string", "null"]
        result:
          type: ["string", "null"]
        learnings:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, analyzing, decided, archived]
        confidenceScore:
          type: integer
          minimum: 0
          maximum: 100
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DecisionInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        context:
          type: string
        options:
          type: array
          items:
            type: string
        advantages:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
    DecisionUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        context:
          type: ["string", "null"]
        options:
          type: array
          items:
            type: string
        advantages:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        result:
          type: ["string", "null"]
        learnings:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, analyzing, decided, archived]
        confidenceScore:
          type: integer
          minimum: 0
          maximum: 100

    # ── ASSETS ────────────────────────────────────────────────────────────────
...
    TasksFromDecisionInput:
      type: object
      required: [tasks]
      properties:
        tasks:
          type: array
          items:
            type: object
            required: [title]
            properties:
              title:
                type: string
              priority:
                type: string
                enum: [low, medium, high, urgent]
...